@@ -11,9 +11,10 @@
         }
     ],
     "require": {
-        "php": ">= 7.1",
+        "php": ">= 8.0",
         "acdh-oeaw/easyrdf": "*",
-        "acdh-oeaw/arche-assets": "*"
+        "acdh-oeaw/arche-assets": "^3.3.1",
+        "guzzlehttp/guzzle": "^7.5"
     },
     "autoload": {
         "psr-4": {
@@ -21,6 +22,7 @@
         }
     },
     "require-dev": {
-        "phpunit/phpunit": "^9"
+        "phpunit/phpunit": "^9",
+        "phpstan/phpstan": "^1.8"
     }
 }
